Replace most of the logging with WINE stuff
Instead of using our own logging thing, I'm not just using WINE's debug
header on the wine side. That makes debugging the program more similar
to debugging dxvk or amf. That seems nice to me. Also it lets me NOT
have all the debug spam.

I'll work on reducing the spam a little more in the future, but being
able to ignore it is at least nice.