Send the current indecies to the real driver
The real vrlink driver doesn't swap textures. The entire
GetNextSwapTextureSet implementation is just a no-op. The expectation
seems to be that doing nothing will keep the same texture around. Since
we're doing stuff in different processes, that's not true for us. To
actually persist the indecies I have to send them over before calling
the function, store them in the newly allocated array, call the real
implementation, and then send the result back. In this case I'm doing
all that work just for the implementation to not change the integer. At
least this is technically correct.

The old implementation didn't do that work, and therefore just returned
whatever happened to be on the stack at those addresses back to the
vrserver. That turns out to have been my rendering bug. Now that I've
fixed that it's correctly rendering in the headset for at least a little
while.

I can't believe all that work was caused by this little thing that
I knew was broken. Good thing I decided to fix it on a whim.